Reshuffled the files inside reports to make them more organised so that its easier to add css to specific parts
